+++ shape/cube.h
@@ -4,6 +4,6 @@
 #include "helper/shared_headers.h"
 #include "helper/vertex.h"
 
-void cube(double h, struct vertex** vertices, long* vertices_size, GLushort** indices, long* indices_size);
+void cube(float h, struct vertex** vertices, long* vertices_size, GLushort** indices, long* indices_size);
 
 #endif
